+++ pyanomaly/core/engine/functions/anopred.py
@@ -18,14 +18,14 @@ from torch.utils.data import DataLoader
 
 from pyanomaly.core.utils import AverageMeter, flow_batch_estimate, tensorboard_vis_images, make_info_message, ParamSet
 # from pyanomaly.datatools.evaluate.utils import psnr_error
-from ..abstract.default_engine import DefaultTrainer, DefaultInference
+from ..abstract.base_engine import BaseTrainer, BaseInference
 
 from ..engine_registry import ENGINE_REGISTRY
 
 __all__ = ['ANOPREDTrainer', 'ANOPREDInference']
 
 @ENGINE_REGISTRY.register()
-class ANOPREDTrainer(DefaultTrainer):
+class ANOPREDTrainer(BaseTrainer):
     NAME = ["ANOPRED.TRAIN"]
     def custom_setup(self):
         # basic meter
@@ -138,7 +138,7 @@ class ANOPREDTrainer(DefaultTrainer):
         self.kwargs['writer_dict']['global_steps_{}'.format(self.kwargs['model_type'])] = global_steps
     
 @ENGINE_REGISTRY.register()
-class ANOPREDInference(DefaultInference):
+class ANOPREDInference(BaseInference):
     NAME = ["ANOPRED.INFERENCE"]    
     def inference(self):
         for h in self._hooks:
