+++ pyanomaly/core/engine/functions/stae.py
@@ -21,44 +21,17 @@ from torch.utils.data import DataLoader
 
 from pyanomaly.core.utils import AverageMeter, flow_batch_estimate, tensorboard_vis_images, make_info_message, ParamSet
 from pyanomaly.datatools.evaluate.utils import psnr_error
-from ..abstract.default_engine import DefaultTrainer, DefaultInference
+from ..abstract.base_engine import BaseTrainer, BaseInference
 from ..engine_registry import ENGINE_REGISTRY
 
 __all__ = ['STAETrainer', 'STAEInference']
 
 @ENGINE_REGISTRY.register()
-class STAETrainer(DefaultTrainer):
+class STAETrainer(BaseTrainer):
     NAME = ["STAE.TRAIN"]
     def custom_setup(self):
-
-        # for item_key in self.model.keys():
-        #     attr_name = str(item_key)
-        #     if self.kwargs['parallel']:
-        #         temp_model = self.data_parallel(self.model[item_key])
-        #     else:
-        #         temp_model = self.model[item_key].cuda()
-        #     self.__setattr__(attr_name, temp_model)
-        
-        # # get the optimizer
-        # for item_key in self.optimizer.keys():
-        #     attr_name = str(item_key)
-        #     # get the optimizer
-        #     self.__setattr__(attr_name, self.optimizer[item_key])
-        #     # get the lr scheduler
-        #     self.__setattr__(f'{attr_name}_scheduler', self.lr_scheduler_dict[f'{attr_name}_scheduler'])
-
-
-        # import ipdb; ipdb.set_trace()
-
-        # get the loss_fucntion
-        # self.rec_loss = self.loss_function['RecLoss']
-        # self.pred_loss = self.loss_function['WeightedPredLoss']
-        
-        
-
         # basic meter
         self.loss_meter_STAE = AverageMeter(name='loss_STAE')
-
 
     def train(self,current_step):
         # Pytorch [N, C, D, H, W]
@@ -132,20 +105,8 @@ class STAETrainer(DefaultTrainer):
         self.kwargs['writer_dict']['global_steps_{}'.format(self.kwargs['model_type'])] = global_steps
     
 @ENGINE_REGISTRY.register()
-class STAEInference(DefaultInference):
+class STAEInference(BaseInference):
     NAME = ["STAE.INFERENCE"]
-    def custom_setup(self, *defaults,**kwargs):
-        # if self.kwargs['parallel']:
-        #     self.STAE = self.data_parallel(self.model['STAE']).load_state_dict(self.save_model['STAE'])
-        # else:
-        #     # import ipdb; ipdb.set_trace()
-        #     self.STAE = self.model['STAE'].cuda()
-        #     self.STAE.load_state_dict(self.save_model['STAE'])
-        
-        # self.test_dataset_keys = self.kwargs['test_dataset_keys']
-        # self.test_dataset_dict = self.kwargs['test_dataset_dict']
-        pass
-
     def inference(self):
         for h in self._hooks:
             h.inference()
